+++ src/TRSConversion/Parse/COPS/Problem.hs
@@ -4,9 +4,11 @@ Description : Parser a Problem in COPS format
 -}
 module TRSConversion.Parse.COPS.Problem (
   parseProblem,
+  parseCOMProblem,
 )
 where
 
+import TRSConversion.Parse.COPS.COM (parseCopsCom)
 import TRSConversion.Parse.COPS.CSCTrs (parseCopsCSCTrs)
 import TRSConversion.Parse.COPS.CSTrs (parseCopsCSTrs)
 import TRSConversion.Parse.COPS.CTrs (parseCopsCTrs)
@@ -14,13 +16,14 @@ import TRSConversion.Parse.COPS.MSTrs (parseCopsMsTrs)
 import TRSConversion.Parse.COPS.MetaInfo (parseCopsMetaInfoBlock)
 import TRSConversion.Parse.COPS.Trs (parseCopsTrs)
 import TRSConversion.Parse.COPS.Utils (COPSParser)
-import TRSConversion.Problem.Common.MetaInfo (emptyMetaInfo)
+import TRSConversion.Problem.Common.MetaInfo (MetaInfo (..), emptyMetaInfo)
 import TRSConversion.Problem.Problem (Problem (Problem))
 import qualified TRSConversion.Problem.Problem as Prob
 import Text.Megaparsec (choice, option, try)
+import Control.Applicative ((<|>))
 
 parseProblem :: COPSParser Problem
-parseProblem = do
+parseProblem = parseCOMProblem <|> do
   system <-
     choice
       [ try $ Prob.Trs <$> parseCopsTrs
@@ -34,4 +37,14 @@ parseProblem = do
     Problem
       { Prob.metaInfo = metaInfo
       , Prob.system = system
+      }
+
+parseCOMProblem :: COPSParser Problem
+parseCOMProblem = do
+  (copsComment, trs) <- parseCopsCom
+  metaInfo <- option emptyMetaInfo parseCopsMetaInfoBlock
+  pure $
+    Problem
+      { Prob.metaInfo = metaInfo{origin = copsComment}
+      , Prob.system = Prob.Trs trs
       }