+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to the
 
 ## Unreleased
 - ARI format supports for multiple systems
-  (not the case yet for COPS)
+- new flag '--comm' for printing COMMUTATION problems in COPS
+- COPS format supports COMMUTATION problems
 
 ## 0.3.0.0 - 2023-05-26
 - restructure Parser modules to separate COPS and ARI parsers